[PATCH] es7000 subarch update
From: "Protasevich, Natalie" <Natalie.Protasevich@UNISYS.com>

The patch fixes a problem with ES7000 Server Management mechanism that uses
platform register mip_port.  It was not initialized, so the mechanism was not
functional.

The patch also fixes the APIC destination for hierarchical and flat cluster
models used in ES7000.  The destination ID's reflect policies for Cascade
based systems which use logical delivery and lowest priority mechanism, and
for xAPIC based models that use physical delivery and fixed APIC destinations.

The patch also turns on NO_IOAPIC_CHECK (1) to avoid error messages and
attempts to re-write the ID, because on ES7000 all ID's are hard coded in the
BIOS and cannot be altered.

BKrev: 409fecacrHdySp6EG1MOm4R5PclOvA