[PATCH] readahead: keep file->f_ra sane
When two threads are simultaneously pread()ing from the same fd (which is a
legitimate thing to do), the readahead code thinks that a huge amount of
seeking is happening and shrinks the window, damaging performance a lot.

I don't see a sane way to avoid this within the readahead code, so take a
private copy of the readahead state and restore it prior to returning from the
read.

BKrev: 409fee8eEgon-0kUEmemNeglYnS1Uw