[PATCH] es7000 subarch update for generic arch
From: "Protasevich, Natalie" <Natalie.Protasevich@UNISYS.com>

This is ES7000 sub architecture update.  It makes ES7000 a part of the
generic architecture, so the single compiled kernel will be able to choose
a correct set of parameters, routines ("genapic"), and a boot path.  It
uses criteria provided by the subarch for platform identification.  In case
of ES7000, it is a unique product/vendor string in the ACPI/MP OEM table,
and server control registers.  The patch is confined to only es7000 subarch
and generic subarch.  It was tested on ES7000 as well as generic Intel 8x
Xeon system.  Andi Kleen has reviewed the changes.

BKrev: 409feee2z290OOV70MQ3WE0uZieejw