[PATCH] fix wrong var used in hotplug/shpchp_ctrl.c.
From: "Luiz Fernando N. Capitulino" <lcapitulino@prefeitura.sp.gov.br>

Zhenmin's checker tool <zli4@cs.uiuc.edu> detected this:

 9. /drivers/pci/hotplug/shpchp_ctrl.c, Line 1575:
 err("%s: Failed to disable slot, error code(%d)\n", __FUNCTION__, rc);

 Maybe change to:
 err("%s: Failed to disable slot, error code(%d)\n", __FUNCTION__,
 retval);

I think it is right because at line 1564, the slot is turned off, and in
this line (1575) is checked the status to see if we got an error; if so,
the error number is shown.  This number is in 'retval', not in 'rc' ('rc'
does have the return of configure_new_device()).

BKrev: 409fef56K9oQXYcJq5gScOI-8hA3RQ