[PATCH] PPC termio fix
From: Paul Mackerras <paulus@samba.org>

It turns out that we are not handling the TABDLY bits of the termios
c_oflag field correctly on PPC, PPC64 and Alpha.  These three architectures
have a value for XTABS that is different from the TAB3 value.  POSIX
specifies that setting the TABDLY field to TAB3 should result in tabs being
expanded to spaces.  In n_tty.c:opost() we check for O_TABDLY(tty) ==
XTABS, which is fine on most architectures because they have XTABS == TAB3.

I think the right thing to do is just to change the definition of XTABS to
be the same as TAB3 on these architectures.  The patch below does this for
PPC and PPC64 (and I suggest the Alpha maintainer should do the same).  At
the moment, applications using either the XTABS or TAB3 values won't get
the expected behaviour.  With this patch, apps that use TAB3 will get the
expected behaviour.  Apps that use XTABS will need to be recompiled (but
note that the POSIX-specified name to use is TAB3 not XTABS).

BKrev: 409ff02a2H3ejnvOTruhd_mIzEN4Vw