Merge redhat.com:/spare/repo/linux-2.6 into redhat.com:/spare/repo/net-drivers-2.6
2004/05/11 03:05:10-04:00 mingo
[netdrvr dmfe] netpoll support

2004/05/11 03:05:01-04:00 hch
[netdrvr gt961000eth] remove useless MOD_{INC,DEC}_USE_COUNT

*grr* - this one slipped in after my last round of audits it seems.

2004/05/11 03:04:53-04:00 akpm
[PATCH] 8139too not running s3 suspend/resume pci fix

From: "Adrian Yee" <brewt-linux-kernel@brewt.org>

Having an 8139 based device in my notebook, I often switch between it and
wireless.  The problem is that the 8139too driver does not save/restore the
pci configuration of the card if the device isn't running.  This simple
patch moves the save/restore code so that the code runs regardless of
whether or not the device is running.

I looked at other drivers and they all seem to do the same thing.  Is there
a reason why this isn't done like in the patch?

2004/05/11 03:04:44-04:00 akpm
[PATCH] lance.c: fix for card with signature 0x52 0x49

From: Vesselin Kostadiov <vesselin@alphawave.com.au>

Problem: The lance.c driver did not work with my Racal Interlan EtherBlaster
card.

More info: I found that my card has a signature 0x52 0x49 that was not
recognized by the driver.

Explanation: Following your suggestion I created a static table with possible
signatures, not too different from the table in ni65.c.  The updated code
compares the first byte of the cards signature with the first byte of the
signatures from the table.  It this succeeds then it reads the second byte and
compares the whole signature with the values from the table.  This way the
minimal I/O reads approach is maintained.

Side effect: The previous version would missdetect cards with signatures 0x52
0x57 and 0x57 0x44.  This has been fixed as well.

2004/05/11 03:04:36-04:00 cw
[PATCH] kill warning in r8169

BKrev: 40a568c9IxqEsyGXG-FKZ11qxr5VQw