[PATCH] dentry d_bucket fix
The gap between checking d_bucket and sampling d_move_count looks like a bug
to me.

It feels safer to be checking d_bucket after taking the lock, when we know
that it is stable.

And it's a little faster to check d_bucket after having checked the hash
rather than before.

BKrev: 40a584aakydPvsR3y9tH3PKqhYaIhQ