[PATCH] ppc64 iSeries: allow read only virtual disks
From: Stephen Rothwell <sfr@canb.auug.org.au>

It is possible to attach a virtual disk to a logical partition on an iSeries
machine so that it is read only to the partition.  This patch allows Linux to
use such virtual disks.

BKrev: 40a586ddVP5tvGG0q95796Dg-_il5g