Merge bk://kernel.bkbits.net/jgarzik/i810-audio-2.6 into ppc970.osdl.org:/home/torvalds/v2.6/linux
2004/05/15 01:03:42-04:00 jgarzik
[sound/oss i810] bump driver to version 1.00

2004/05/15 00:18:55-04:00 jgarzik
Merge redhat.com:/spare/repo/linux-2.6
into redhat.com:/spare/repo/i810-audio-2.6

2004/05/11 13:31:42-04:00 gtw
[sound i810] silently ignore invalid PCM_ENABLE_xxx bits from userland

We must guarantee that struct file's ->f_mode agrees with
PCM_ENABLE_xxx bits from userland OSS apps.  Other drivers silently
ignore invalid bits, so we follow their lead.

2004/05/11 04:05:08-04:00 herbert
[sound/oss i810] fix deadlock in drain_dac

This patch fixes a typo in a previous change that causes the driver
to deadlock under SMP.

2004/05/11 04:04:09-04:00 herbert
[sound/oss i810] fix reads/writes % 4 != 0

This patch removes another bogus chunk of code that breaks when
the application does a partial write.

In particular, a read/write of x bytes where x % 4 != 0 will loop forever.

2004/05/11 04:02:55-04:00 herbert
[sound/oss i810] fix drain_dac loop when signals_allowed==0

This patch fixes another bug in the drain_dac wait loop when it is
called with signals_allowed == 0.

2004/05/11 04:01:34-04:00 herbert
[sound/oss i810] remove divides on playback

This patch removes a couple of divides on the playback path.

2004/05/11 03:59:23-04:00 herbert
[sound/oss i810] fix OSS fragments

This patch makes userfragsize do what it's meant to do: do not start
DAC/ADC until a full fragment is available.

2004/05/11 03:58:35-04:00 herbert
[sound/oss i810] fix playback SETTRIGGER

This patch fixes SETTRIGGER with playback so that the LVI is always
set and the DAC is always started.

2004/05/11 03:57:31-04:00 herbert
[sound/oss i810] fix partial DMA transfers

This patch fixes a longstanding bug in this driver where partial fragments
are fed to the hardware.  Worse yet, those fragments are then extended
while the hardware is doing DMA transfers causing all sorts of problems.

2004/05/11 03:56:34-04:00 herbert
[sound/oss i810] clean up with macros

This patch adds a number macros to clean up the code.

2004/05/11 03:52:23-04:00 herbert
[sound/oss] remove bogus CIV_TO_LVI

This patch removes a pair of bogus LVI assignments.  The explanation in
the comment is wrong because the value of PCIB tells the hardware that
the DMA buffer can be processed even if LVI == CIV.

Setting LVI to CIV + 1 causes overruns when with short writes
(something that vmware is very fond of).

2004/05/11 03:50:39-04:00 herbert
[sound/oss i810] fix race

This patch fixes the value of swptr in case of an underrun/overrun.

Overruns/underruns probably won't occur at all when the driver is
fixed properly, but this doesn't hurt.

2004/05/11 03:48:33-04:00 herbert
[sound/oss i810] fix wait queue race in drain_dac

This particular one fixes a textbook race condition in drain_dac
that causes it to timeout when it shouldn't.

BKrev: 40a5a7d0pEH1-eZEn0pwThpGDS0BuA