[PATCH] x86-64: fix /dev/mem caching behaviour
This changes the /dev/mem caching behaviour on x86-64 to be compatible
with i386.

By default everything is set cached.

This actually makes WC MTRRs on AMD systems work, which would get
overriden by the UC PAT bits that were set earlier.  This can make DVD
decoding with hardware support a lot faster.

It also supports O_SYNC now, like i386, although that is not really
safe, because it allows the user to create undefined cache attribute
conflicts that can corrupt caches in some circumstances.  I kept it for
now.  Better would to disallow it, until Terrence Ripperda's PAT
framework is getting merged, that can avoid these problems.

Actually it would be probably a good idea to add a printk here to catch
broken programs for i386 and x86-64, but that is for another patch.

BKrev: 40a6562faKfIkf9klbU2bAjlfh1xMA