Merge bk://gkernel.bkbits.net/libata-2.6 into ppc970.osdl.org:/home/torvalds/v2.6/linux
2004/05/16 18:48:08-04:00 jgarzik
[libata] remove redundant use of ATA_QCFLAG_SG in ATAPI packet translation

ata_scsi_translate() sets this flag for all ATA->SCSI translated
commands, so it need not be done in atapi_xlat().  The now-removed
use in atapi_xlat() was also inconsistent WRT PIO versus DMA.

2004/05/16 18:35:15-04:00 jgarzik
[libata] DMADIR support

DMADIR bit is necessary for some PATA->SATA bridges.  These
bridges require the OS driver to specify the data xfer direction,
for PACKET (a.k.a. scsi) commands.

A reliable DMADIR detection method hasn't yet been developed,
and ATAPI is still a WIP, so DMADIR is enabled with an ifdef for
now.

2004/05/15 18:07:32-04:00 jgarzik
[libata] handle non-data ATAPI commands via interrupt

It's easier to do it this way, than polling, at the moment.

Also, fix a test in ata_scsi_translate that was incorrectly
erroring-out non-data commands.

BKrev: 40a827d9ndFgugqbGlJhy_hHjm9Dqg