[PATCH] alpha fp-emu vs module refcounting
From: Ivan Kokshaysky <ink@jurassic.park.msu.ru>

This allows building the math-emu code as a module only when
CONFIG_SMP is not set. The fp trap handler cannot be preempted
on a single-CPU (as CONFIG_PREEMPT is not going to be supported
on alpha), so the module can be safely unloaded at any time.

BKrev: 40ab8d25TlElMtMYC_HKlxylsseCgA