[PATCH] PPC32: Get full register set on bad kernel accesses
From: Paul Mackerras <paulus@samba.org>

At present on ppc32, if the kernel accesses a bad address and causes an
oops, or drops into the xmon debugger, we only have the contents of the
volatile registers available to print.  The reason is that we only save the
volatile registers on entry for a page fault.

This patch restructures the code a bit so that if do_page_fault()
determines that the page fault is caused by a bad kernel access, it returns
to the caller, which then saves the full register set into the exception
frame before calling bad_page_fault().  This way we get the full set of
registers printed in the oops message.

BKrev: 40abeeb4NEyIkqKvB8G_2LPZhtx1SA