[PATCH] dentry size tuning
Experimenting with various values of DENTRY_STORAGE

 dentry size      objs/slab   dentry size * objs/slab    inline string

   148               26               3848                   32
   152               26               3952                   36
   156               25               3900                   40
   160               24               4000                   44

We're currently at 160.  The patch fairly arbitrarily takes it down to 152, so
we can fit a 35-char name into the inline part of the dentry.

Also, go back to the old way of sizing d_iname so that any arch-specific
compiler-forced alignemnts are honoured.

BKrev: 40abef2aXwrTlMXgXuqEOJ_pTJJjQg