[PATCH] replace MOD_INC_USE_COUNT in cyber2000fb
From: Christoph Hellwig <hch@lst.de>

This driver is unloadable for the pci case, but not if vlb cards are found so
we can't use the module_exit removal to lock it into memory.

Replace the MOD_INC_USE_COUNT with __module_get in it's module_init routine.

BKrev: 40abefa2d2VKtOGjGtz2LgwuJd3MMw