[PATCH] Invalid notify_change(symlink, [ATTR_MODE]) in nfsd
Make sure NFS client doesn't see errors from mode setting on new symlinks.

When nfsd creates a symlink, it tries to set the mode as the mode is
carried in the NFS request and some filesystems store a mode.

If the filesystem refuses to set the mode (e.g. -EOPNOTSUPP), this
error should not be returned to the client.

BKrev: 40accea6R-SBTQ9ijTsYDGEET7CnBg