[PATCH] ppc64: make enter_rtas() take unsigned long arg
From: Paul Mackerras <paulus@samba.org>

We declare enter_rtas with a struct rtas_args * argument, though it is
supposed to be a physical address, and then every time we call it we cast the
unsigned long result from __pa() to a void *.  This patch changes the
declaration of enter_rtas to make it take an unsigned long argument, and
removes the cast from all the callers.  The actual enter_rtas() routine is in
assembler and doesn't need to be changed.

BKrev: 40ae6424Vlmd6Y1xRPdF2M5yZ3yzyQ