[PATCH] ramdisk: fix PageUptodate() handling
When a filesystem does getblk() to get a buffer_head against the ramdisk the
VFS will allocate a new not-uptodate pagecache page and will attach buffers to
it.

The filesystem will then bring certain buffer_heads uptodate.  But not the
whole page.

Later, various ramdisk a_ops see the not-uptodate page and wipe the whole
thing out, including the parts to which the filesystem wrote!

Fix that up by only zapping those parts of the page which are covered by
non-uptodate buffers.

BKrev: 40ae645c1_pOBsZ2iFDwID9_YS7zOQ