[PATCH] drop left-over #ifndef __ia64__
From: David Mosberger <davidm@napali.hpl.hp.com>

It used to be that loops_per_jiffy was a macro on ia64, hence it couldn't be
exported.  That's no longer the case though, so there is no point in
inhibiting its export (not that it makes any _sense_ to export that value on
ia64).

BKrev: 40ae655beNE4qAs6TBmp4iqaRZ-bOA